Support projects who don't commit stack.yaml.lock
When not present, the `stack.yaml` itself will be used as the cache key.
This only matters in the edge-case of transitive dependencies through
the named resolver changing revisions or hashes without the resolver
name itself changing.

It's unclear when that may happen, but if it does, a false-positive
cache hit is probably not problematic. Folks who care about correctness
to this degree, will be committing the `.lock` file for its
reproducibility guarantees anyway.

Fixes #68.