Ensure ./.stack-work is always included
Without this, we see partial caching in a multi-package project.

The example in this project is a little un-realistic. It contains a
top-level package and a sub-directory package. Typically, there are only
sub-directory packages. For such cases, this action would've omitted
./.stack-work and not cached correctly. This fixes that and ensures it's
included for simple projects, top-and-sub projects (like example), and
sub-projects (as is common).

It's unclear if we need to unique, but I'm doing so out of caution and
to at least prevent confusion when the values are logged.